backend/requirements/dev.txt
-r base.txt
ruff>=0.4
pytest>=8.0
pytest-django>=4.8
